requirements.txt
arrow==0.12.1
dj_database_url==0.5.0
django-emoticons==1.2
django-forms-bootstrap==3.1.0
django-tinymce==2.7.0
gunicorn==19.9.0
stripe==2.0.1